Standard types

+++ Snippets/08. sieve/08. sieve.cpp
@@ -11,13 +11,13 @@
 
 /* ******************************************************************************** */
 
-int sieve(int size){
+i32 sieve(i32 size){
     assert(size>2);
 
-    int i;
-    int j;
+    i32 i;
+    i32 j;
     
-    std::vector<int> vec(size);
+    std::vector<i32> vec(size);
 
     for(vec[1]=0, i=2; i<size; i++)
         vec[i]=1;
@@ -27,7 +27,7 @@ int sieve(int size){
             if((i*j)<size)
                 vec[i*j] = 0;
 
-    int biggestIdx = 0;
+    i32 biggestIdx = 0;
 
     for(i = 1; i < size; i++)
         if(vec[i])
@@ -39,14 +39,14 @@ int sieve(int size){
 /* ******************************************************************************** */
 
 sieveResult sieveUnbounded(){
-    unsigned long long i;
-    unsigned long long j;
+    size_t i;
+    size_t j;
 
     std::chrono::steady_clock::time_point begin = std::chrono::steady_clock::now();
 
     size_t size = 30000000;
 
-    std::vector<int> vec(pow(10, 9));
+    std::vector<i32> vec(pow(10, 9));
 
     for(vec[1] = 0, i = 2; i < size; i++)
         vec[i] = 1;
@@ -56,7 +56,7 @@ sieveResult sieveUnbounded(){
             if((i * j) < size)
                 vec[i * j] = 0;
 
-    int biggestIdx = 0;
+    i32 biggestIdx = 0;
 
     for(i = 1; i < size; i++)
         if(vec[i])
@@ -64,7 +64,7 @@ sieveResult sieveUnbounded(){
 
     std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();
 
-    double d = std::chrono::duration_cast<std::chrono::microseconds>(end - begin).count() / 1000000.0;
+    f64 d = std::chrono::duration_cast<std::chrono::microseconds>(end - begin).count() / 1000000.0;
 
     sieveResult result;
     result.idx = biggestIdx;
@@ -76,8 +76,8 @@ sieveResult sieveUnbounded(){
 /* ******************************************************************************** */
 
 sieveResult sieveOptimized(){
-    unsigned long long i;
-    unsigned long long j;
+    size_t i;
+    size_t j;
 
     sieveResult res;
 
@@ -85,16 +85,16 @@ sieveResult sieveOptimized(){
 
     size_t n = 340000000;
 
-    std::vector<int> lp(n+1);
-    std::vector<int> pr;
+    std::vector<i32> lp(n+1);
+    std::vector<i32> pr;
 
-    for (int i=2; i <= n; ++i) {
+    for (i32 i=2; i <= n; ++i) {
         if (lp[i] == 0) {
             lp[i] = i;
             pr.push_back(i);
         }
 
-        for (int j = 0; i * pr[j] <= n; ++j) {
+        for (i32 j = 0; i * pr[j] <= n; ++j) {
             lp[i * pr[j]] = pr[j];
             if (pr[j] == lp[i]) {
                 break;
@@ -102,7 +102,7 @@ sieveResult sieveOptimized(){
         }
     }
 
-    int biggestIdx = 0;
+    i32 biggestIdx = 0;
 
     for(i = 1; i < n; i++)
         if(lp[i])
@@ -110,7 +110,7 @@ sieveResult sieveOptimized(){
 
     std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();
 
-    double d = std::chrono::duration_cast<std::chrono::microseconds>(end - begin).count() / 1000000.0;
+    f64 d = std::chrono::duration_cast<std::chrono::microseconds>(end - begin).count() / 1000000.0;
     res.idx = biggestIdx;
     res.seconds = d;
